+++ component/frontend/ViewTemplates/Item/form.blade.php
@@ -34,6 +34,7 @@ $captcha = $this->getCaptchaField();
                 'filter_order_Dir' => 'asc',
                 'limit' => 100,
                 'limitstart' => 0,
+                'enabled' => 1,
             ]) }}
         </div>