+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * gh-15 Akismet support 
 * Set default backend Items ordering to ID descending (new messages first)
 * Common PHP version warning scripts
+* Dark Mode
 
 # 2.0.1